Extract AbstractSpscLinkedArrayQueue base class.
Factor out SpscGrowableArrayQueue and SpscUnboundedArrayQueue shared code.
